requirements.txt
streamlit>=1.32.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
pypdf>=4.0.0
python-docx>=1.1.0
requests>=2.31.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.38.0
